+++ src/commands/openFileRevisionFrom.ts
@@ -1,46 +1,43 @@
 'use strict';
 import { Range, TextDocumentShowOptions, TextEditor, Uri } from 'vscode';
 import { GlyphChars } from '../constants';
-import { GitBranch, GitTag, GitUri } from '../git/gitService';
-import { BranchesAndTagsQuickPick, BranchQuickPickItem, TagQuickPickItem } from '../quickpicks';
+import { GitBranch, GitReference, GitTag, GitUri } from '../git/gitService';
+import { CommandQuickPickItem, ReferencesQuickPick } from '../quickpicks';
 import { Strings } from '../system';
 import { ActiveEditorCommand, command, Commands, getCommandUri, openEditor } from './common';
 
-export interface OpenFileRevisionFromBranchCommandArgs {
-    branchOrTag?: GitBranch | GitTag;
+export interface OpenFileRevisionFromCommandArgs {
+    reference?: GitBranch | GitTag | GitReference;
 
     line?: number;
     showOptions?: TextDocumentShowOptions;
 }
 
 @command()
-export class OpenFileRevisionFromBranchCommand extends ActiveEditorCommand {
+export class OpenFileRevisionFromCommand extends ActiveEditorCommand {
     constructor() {
-        super(Commands.OpenFileRevisionFromBranch);
+        super([Commands.OpenFileRevisionFrom, Commands.OpenFileRevisionFromBranch]);
     }
 
-    async execute(editor: TextEditor | undefined, uri?: Uri, args: OpenFileRevisionFromBranchCommandArgs = {}) {
+    async execute(editor: TextEditor | undefined, uri?: Uri, args: OpenFileRevisionFromCommandArgs = {}) {
         uri = getCommandUri(uri, editor);
         if (uri == null) return undefined;
 
         const gitUri = await GitUri.fromUri(uri);
         if (!gitUri.repoPath) return undefined;
 
-        if (args.branchOrTag === undefined) {
+        if (args.reference === undefined) {
             const placeHolder = `Open revision of ${gitUri.getFormattedPath()}${
                 gitUri.sha ? ` ${Strings.pad(GlyphChars.Dot, 1, 1)} ${gitUri.shortSha}` : ''
-            } from Branch or Tag${GlyphChars.Ellipsis}`;
+            } from${GlyphChars.Ellipsis}`;
 
-            const pick = await new BranchesAndTagsQuickPick(gitUri.repoPath).show(placeHolder, {
-                allowCommitId: false
+            const pick = await new ReferencesQuickPick(gitUri.repoPath).show(placeHolder, {
+                allowEnteringRefs: true
             });
             if (pick === undefined) return undefined;
+            if (pick instanceof CommandQuickPickItem) return pick.execute();
 
-            if (!(pick instanceof BranchQuickPickItem) && !(pick instanceof TagQuickPickItem)) {
-                return undefined;
-            }
-
-            args.branchOrTag = pick.item;
+            args.reference = pick.item;
         }
 
         if (args.line !== undefined && args.line !== 0) {
@@ -50,7 +47,7 @@ export class OpenFileRevisionFromBranchCommand extends ActiveEditorCommand {
             args.showOptions.selection = new Range(args.line, 0, args.line, 0);
         }
 
-        return openEditor(GitUri.toRevisionUri(args.branchOrTag.ref, gitUri.fsPath, gitUri.repoPath), {
+        return openEditor(GitUri.toRevisionUri(args.reference.ref, gitUri.fsPath, gitUri.repoPath), {
             ...args.showOptions,
             rethrow: true
         });