+++ src/quickpicks.ts
@@ -1,6 +1,5 @@
 'use strict';
 
-export * from './quickpicks/branchesAndTagsQuickPick';
 export * from './quickpicks/branchHistoryQuickPick';
 export * from './quickpicks/commitFileQuickPick';
 export * from './quickpicks/commitQuickPick';
@@ -8,6 +7,7 @@ export * from './quickpicks/commitsQuickPick';
 export * from './quickpicks/commonQuickPicks';
 export * from './quickpicks/fileHistoryQuickPick';
 export * from './quickpicks/modesQuickPick';
+export * from './quickpicks/referencesQuickPick';
 export * from './quickpicks/remotesQuickPick';
 export * from './quickpicks/repositoriesQuickPick';
 export * from './quickpicks/repoStatusQuickPick';